+++ SourceryTests/Parsing/ComposerSpec.swift
@@ -26,6 +26,11 @@ class ParserComposerSpec: QuickSpec {
                     return Composer.uniqueTypesAndFunctions(parserResult).types
                 }
 
+                func parseFunctions(_ code: String) -> [SourceryMethod] {
+                    guard let parserResult = try? FileParser(contents: code).parse() else { fail(); return [] }
+                    return Composer.uniqueTypesAndFunctions(parserResult).functions
+                }
+
                 context("given class hierarchy") {
                     var fooType: Type!
                     var barType: Type!
@@ -993,6 +998,46 @@ class ParserComposerSpec: QuickSpec {
                         expect(types).to(equal([expectedBar, expectedFoo]))
                         expect(types.last?.variables.first?.type).to(equal(expectedBar))
                         expect(types.last?.variables.first?.definedInType).to(equal(expectedFoo))
+                    }
+                }
+
+                context("given free function") {
+                    it("resolves generic return types properly") {
+                        let functions = parseFunctions("func foo() -> Bar<String> { }")
+                        expect(functions[0]).to(equal(SourceryMethod(
+                            name: "foo()",
+                            selectorName: "foo",
+                            parameters: [],
+                            returnTypeName: TypeName(
+                                "Bar<String>",
+                                generic: GenericType(
+                                    name: "Bar",
+                                    typeParameters: [
+                                        GenericTypeParameter(
+                                            typeName: TypeName("String"),
+                                            type: nil
+                                        )
+                                ])
+                            ),
+                            definedInTypeName: nil)))
+                    }
+
+                    it("resolves tuple return types properly") {
+                        let functions = parseFunctions("func foo() -> (bar: String, biz: Int) { }")
+                        expect(functions[0]).to(equal(SourceryMethod(
+                            name: "foo()",
+                            selectorName: "foo",
+                            parameters: [],
+                            returnTypeName: TypeName(
+                                "(bar: String, biz: Int)",
+                                tuple: TupleType(
+                                    name: "(bar: String, biz: Int)",
+                                    elements: [
+                                        TupleElement(name: "bar", typeName: TypeName("String")),
+                                        TupleElement(name: "biz", typeName: TypeName("Int"))
+                                    ])
+                            ),
+                            definedInTypeName: nil)))
                     }
                 }
             }