+++ SourceryTests/Parsing/FileParser + MethodsSpec.swift
@@ -20,6 +20,11 @@ class FileParserMethodsSpec: QuickSpec {
                     return Composer.uniqueTypesAndFunctions(parserResult).types
                 }
 
+                func parseFunctions(_ code: String) -> [SourceryMethod] {
+                    guard let parserResult = try? FileParser(contents: code).parse() else { fail(); return [] }
+                    return Composer.uniqueTypesAndFunctions(parserResult).functions
+                }
+
                 it("extracts methods properly") {
                     let methods = parse("""
                     class Foo {
@@ -109,6 +114,12 @@ class FileParserMethodsSpec: QuickSpec {
                             Method(name: "foo()", selectorName: "foo", isStatic: true, definedInTypeName: TypeName("Foo"))
                             ])
                         ]))
+                }
+
+                it("extracts free functions properly") {
+                    expect(parseFunctions("func foo() {}")).to(equal([
+                        Method(name: "foo()", selectorName: "foo", isStatic: false, definedInTypeName: nil)
+                    ]))
                 }
 
                 context("given method with parameters") {